Start to modularise the rust code a bit
